New kafka-consumer-groups functionality
This release changes from the old method of using 'bootstrap' server.
That approach meant that any user of the library would likely need
to customize the python code in order to get their environment working.

Th new method is to specify the full kafka-consumer-groups.sh commands
for '--list' and '--describe' as an array in consumerlag.yaml.

Another improvement is to allow testing of Kafka interaction without
interaction with Dynatrace.  This allows a user of this library to
test the script against their Kafka cluster in separation from
their Dynatrace environment.  This new flag is called 'kafka_only'
and is set to True in the supplied consumerlag.yaml file.