Add split_request functionality
This release allows the ability to configure a maximum byte size per request.
The default value will be 10000 bytes unless it is specified as
'send_byte_size_limit' within consumerlag.yaml as an alternate value.

Logging will also explain why and how the splitting is occurring.

See log entries for 'SplitRequest' to see this information.